+++ frontend/src/modules/purchasing/views/PurchasePlanView.tsx
@@ -1,0 +1,8 @@
+import { memo } from "react";
+import PurchasingPlanPage from "../../../pages/purchasing/PurchasingPlanPage";
+
+function PurchasePlanViewInner() {
+  return <PurchasingPlanPage />;
+}
+
+export default memo(PurchasePlanViewInner);